+++ core/dbt/config/renderer.py
@@ -147,7 +147,7 @@ class DbtProjectYamlRenderer(BaseRenderer):
 
         if first in {'seeds', 'models', 'snapshots', 'tests'}:
             keypath_parts = {
-                (k.lstrip('+') if isinstance(k, str) else k)
+                (k.lstrip('+ ') if isinstance(k, str) else k)
                 for k in keypath
             }
             # model-level hooks